% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.markovMSM.R
\name{plot.markovMSM}
\alias{plot.markovMSM}
\title{Plot for an object of class "markovMSM"}
\usage{
\method{plot}{markovMSM}(
  x = object,
  y = NULL,
  quantileOrder = NULL,
  tran = "pstate2",
  axis.scale = c(-1, 1),
  difP = "False",
  ...
)
}
\arguments{
\item{x}{A dataframe in long format in case of the plot with the differences between AJ and LMAJ estimators
or a object of class "markovMSM" with the results of the global.test or local.test function.}

\item{y}{Parameter with the value NULL.}

\item{quantileOrder}{Order of the quantil used in the formula of the AUC global test.}

\item{tran}{Transition probability from the starting state. Example: if objet 'from' of global is equal
1 then tran='pstate2' means the results of the plot correspond to the difference between AJ and LMAJ
estimates of transition probabilites of the transition 1->2.}

\item{axis.scale}{Limit of the y axis of the plots.}

\item{difP}{Type of plot representing in case of x means the results of the Localtest. If difP='True' plot 
depicts the discrepancies between AJ and LMAJ estimators.  If difP='False' plot show the AJ and 
the LMAJ estimates. By default difP='False'.}

\item{...}{For future methods.}
}
\value{
No value is returned.
}
\description{
Plots for the visual inspection for checking the Markov condition through the difference 
between the AJ and LMAJ from a starting time of the transition probabilities for each transition.
}
\examples{
 
library(mstate)

data(prothr)

res<-local.test(db_long=prothr, times=30, from=1, to=3)

plot(x=res, tran='pstate2', axis.scale=c(-0.25,.25), difP='True')

#plot(x=res, tran='pstate2', axis.scale=c(0,.25), difP='False')

res2<-global.test(db_long=prothr, db_wide = NULL, from=1, to=3, replicas=5, limit=0.90, 
                 quantiles=c(.05, .10, .20, .30, 0.40))

#plot(x=res, quantileOrder=3, tran='pstate2', axis.scale=c(-0.05,.15))

}
\author{
Gustavo Soutinho and Luis Meira-Machado.
}
