% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.markovMSM.R
\name{plot.markovMSM}
\alias{plot.markovMSM}
\title{Plot for an object of class "markovMSM"}
\usage{
\method{plot}{markovMSM}(x, quantileOrder, tran = "pstate2", axis.scale = c(-1, 1), ...)
}
\arguments{
\item{x}{A dataframe in long format in case of the plot with the differences between AJ and LMAJ estimators
or a object of class "markovMSM" with the results of the AUC globalTest function.}

\item{quantileOrder}{Order of the quantil used in the formula of the AUC global test.}

\item{tran}{Transition probability from the starting state. Example: if objet 'from' of global is equal
1 then tran='pstate2' means the results of the plot correspond to the difference between AJ and LMAJ
estimates of transition probabilites of the transition 1->2.}

\item{axis.scale}{Limit of the y axis of the plots.}

\item{...}{For future methods.}
}
\value{
No value is returned.
}
\description{
Plots for the visual inspection for checking the Markov condition through the difference 
between the AJ and LMAJ from a starting time of the transition probabilities for each transition.
}
\examples{

data("ebmt4")
db_wide <- ebmt4
positions=list(c(2, 3, 5, 6), c(4, 5, 6), c(4, 5, 6), c(5, 6), c(6), c())
namesStates =  c("Tx", "Rec", "AE", "Rec+AE", "Rel", "Death")
tmat <-transMatMSM(positions, namesStates)
timesNames = c(NA, "rec", "ae","recae", "rel", "srv")
status=c(NA, "rec.s", "ae.s", "recae.s","rel.s", "srv.s")
trans = tmat
db_long<- prepMSM(data=db_wide, trans, timesNames, status)
res2<-globalTestAUC(db_long, db_wide, from=2, to=3, quantiles=c(.05, .10, .20, .30, 0.40), 
                tmat = tmat, replicas = 20, positions=positions, namesStates=namesStates,
                timesNames=timesNames,status=status)

rm(res)
res<-localTestAUC(db_long, db_wide, times=30, from=3, to=5, tmat = tmat, 
                replicas = 20, namesStates=namesStates, positions=positions,
                timesNames=timesNames,status=status)
plot(res, tran='pstate2', axis.scale=c(0,.2))

}
\author{
Gustavo Soutinho and Luis Meira-Machado.
}
